requirements.txt
flax==0.7.4  
brax==0.9.1 
optax 
trimesh 
gym 
chex==0.1.7 
protobuf==4.24.3 
google 
tensorflow-probability==0.23.0 
dm_env 
matplotlib  
jaxopt 
mujoco 
jinja2 
jraph 
jax-md  
dm-haiku==0.0.10 
e3nn_jax==0.20.2 
scipy==1.10.1
